refactor(output): split activity-handles tests into spinner + progress files
